+++ examples/minimal-consumer/CHANGELOG.md
@@ -1,5 +1,12 @@
 # minimal-consumer
 
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @pwrdrvr/agent-client@0.3.0
+
 ## 0.0.1
 
 ### Patch Changes